html_brief.py
from bs4 import BeautifulSoup, Comment
import json
from utils_logging import log_event

def create_brief(html_content: str) -> dict:
    """
    Create a comprehensive brief from HTML.
    
    Strategy: PAY FOR ACCURACY
    - Extract JSON-LD (structured data gold mine)
    - Send FULL cleaned HTML (no truncation, no guessing)
    - Let the LLM see everything to make accurate decisions
    
    Cost: ~10-15K tokens per page = $0.02-0.04 in API calls
    Value: 95%+ extraction accuracy on first pass
    """
    log_event("Generating HTML Brief (Full Clean Mode)...")
    
    soup = BeautifulSoup(html_content, "lxml")

    # --- STEP 1: EXTRACT JSON-LD FIRST (Before cleaning) ---
    json_ld_data = []
    scripts = soup.find_all("script", type="application/ld+json")
    
    for s in scripts:
        try:
            content = s.string
            if content:
                data = json.loads(content.strip())
                json_ld_data.append(data)
        except:
            continue
    
    log_event(f"Found {len(json_ld_data)} JSON-LD blobs")
            
    # --- STEP 2: CLEAN THE NOISE ---
    # Remove bloat but keep ALL content structure
    removed_tags = 0
    
    # Remove script, style, and other non-content tags
    for tag in soup(["script", "style", "svg", "path", "noscript", "iframe", "meta", "link"]):
        tag.decompose()
        removed_tags += 1
    
    # Remove HTML comments
    for element in soup(text=lambda text: isinstance(text, Comment)):
        element.extract()
    
    log_event(f"Removed {removed_tags} non-content tags")

    # --- STEP 3: BASE64 STRIPPING ---
    # Base64 images/fonts are token killers (100KB+ of gibberish)
    base64_removed = 0
    
    for img in soup.find_all("img"):
        src = img.get("src", "")
        if src.startswith("data:image"):
            img["src"] = "[BASE64_REMOVED]"
            base64_removed += 1
        # Keep all other attributes - they might be useful
    
    log_event(f"Removed {base64_removed} Base64 images")

    # --- STEP 4: GENERATE FULL CLEAN HTML ---
    # NO TRUNCATION - send everything
    # The LLM needs to see the whole structure to be accurate
    clean_html = soup.body.prettify() if soup.body else soup.prettify()
    
    original_size_kb = len(html_content) / 1024
    cleaned_size_kb = len(clean_html) / 1024
    reduction_pct = ((original_size_kb - cleaned_size_kb) / original_size_kb) * 100
    
    log_event(f"Brief complete: {original_size_kb:.1f}KB → {cleaned_size_kb:.1f}KB ({reduction_pct:.1f}% reduction)")
    
    # Estimated tokens (rough: 1 token ≈ 4 chars)
    estimated_tokens = len(clean_html) // 4
    log_event(f"Estimated tokens: ~{estimated_tokens:,}")
    
    brief = {
        "json_ld_data": json_ld_data,        # Structured data (if present)
        "full_clean_html": clean_html,       # Complete cleaned HTML
        "original_size_kb": original_size_kb,
        "cleaned_size_kb": cleaned_size_kb,
        "estimated_tokens": estimated_tokens,
        "base64_images_removed": base64_removed
    }
    
    return brief