requirements.txt
openai>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
jsonpath-ng>=1.6.0
watchdog>=3.0.0
python-dotenv>=1.0.0